requirements.txt
#For web app
chainlit==2.0.1
#For goolge slides
google_api_python_client==2.123.0
google_auth_oauthlib==1.1.0
#For chatbot and read PDF
langchain==0.3.14
langchain_community==0.3.14
langchain_core==0.3.29
langchain_openai==0.3.0
openai==1.59.6

protobuf==5.29.3
